+++ ctakesclient/exceptions.py
